создан контроллер для отрисовки страницы
